#486: Separate main models into separate files
