+++ deviceadmin/__init__.py
@@ -1,0 +1,1 @@
+DEVICE_RESULTS_PER_PAGE = 10
