+++ oppia/models/__init__.py
@@ -1,0 +1,5 @@
+from main import *
+from tag import *
+from badges import *
+from points import *
+from cohorts import *
