+++ oppia/models/main.py
@@ -93,6 +93,7 @@ class Course(models.Model):
             return False
 
     def get_tags(self):
+        from oppia.models import Tag
         tags = Tag.objects.filter(coursetag__course=self)
         str = ""
         for t in tags:
@@ -142,15 +143,6 @@ class Course(models.Model):
         return Tracker.objects.filter(course=course, user=user, completed=True, digest__in=acts).values_list('digest').distinct().count()
 
     @staticmethod
-    def get_points(course, user):
-        points = Points.objects.filter(course=course, user=user).aggregate(total=Sum('points'))
-        return points['total']
-
-    @staticmethod
-    def get_badges(course, user):
-        return Award.objects.filter(user=user, awardcourse__course=course).count()
-
-    @staticmethod
     def get_media_viewed(course, user):
         acts = Media.objects.filter(course=course).values_list('digest')
         return Tracker.objects.filter(course=course, user=user, digest__in=acts).values_list('digest').distinct().count()
@@ -164,32 +156,6 @@ class CourseManager(models.Model):
         verbose_name = _('Course Manager')
         verbose_name_plural = _('Course Managers')
 
-
-class Tag(models.Model):
-    name = models.TextField(blank=False)
-    created_date = models.DateTimeField('date created', default=timezone.now)
-    created_by = models.ForeignKey(User, null=True, on_delete=models.SET_NULL)
-    courses = models.ManyToManyField(Course, through='CourseTag')
-    description = models.TextField(blank=True, null=True, default=None)
-    order_priority = models.IntegerField(default=0)
-    highlight = models.BooleanField(default=False)
-    icon = models.FileField(upload_to="tags", null=True, blank=True, default=None)
-
-    class Meta:
-        verbose_name = _('Tag')
-        verbose_name_plural = _('Tags')
-
-    def __unicode__(self):
-        return self.name
-
-
-class CourseTag(models.Model):
-    course = models.ForeignKey(Course, on_delete=models.CASCADE)
-    tag = models.ForeignKey(Tag, on_delete=models.CASCADE)
-
-    class Meta:
-        verbose_name = _('Course Tag')
-        verbose_name_plural = _('Course Tags')
 
 
 class Section(models.Model):
@@ -475,253 +441,3 @@ class Tracker(models.Model):
 
         if 'lang' in json_data:
             return json_data['lang']
-
- 
-class Cohort(models.Model):
-    description = models.CharField(max_length=100)
-    start_date = models.DateTimeField(default=timezone.now)
-    end_date = models.DateTimeField(default=timezone.now)
-
-    class Meta:
-        verbose_name = _('Cohort')
-        verbose_name_plural = _('Cohorts')
-
-    def __unicode__(self):
-        return self.description
-
-    def no_student_members(self):
-        return Participant.objects.filter(cohort=self, role=Participant.STUDENT).count()
-
-    def no_teacher_members(self):
-        return Participant.objects.filter(cohort=self, role=Participant.TEACHER).count()
-    
-    @staticmethod
-    def student_member_now(course, user):
-        now = timezone.now()
-        cohorts = Cohort.objects.filter(coursecohort__course=course, start_date__lte=now, end_date__gte=now)
-        for c in cohorts:
-            participants = c.participant_set.filter(user=user, role=Participant.STUDENT)
-            for p in participants:
-                return c
-        return None
-
-    @staticmethod
-    def teacher_member_now(course, user):
-        now = timezone.now()
-        cohorts = Cohort.objects.filter(coursecohort__course=course, start_date__lte=now, end_date__gte=now)
-        for c in cohorts:
-            participants = c.participant_set.filter(user=user, role=Participant.TEACHER)
-            for p in participants:
-                return c
-        return None
-
-    @staticmethod
-    def member_now(course, user):
-        now = timezone.now()
-        cohorts = Cohort.objects.filter(coursecohort__course=course, start_date__lte=now, end_date__gte=now)
-        for c in cohorts:
-            participants = c.participant_set.filter(user=user)
-            for p in participants:
-                return c
-        return None
-
-    def get_courses(self):
-        courses = Course.objects.filter(coursecohort__cohort=self).order_by('title')
-        return courses
-
-    def get_leaderboard(self, count=0):
-        users = User.objects.filter(participant__cohort=self,
-                                    participant__role=Participant.STUDENT,
-                                    points__course__coursecohort__cohort=self) \
-                            .annotate(total=Sum('points__points')) \
-                            .order_by('-total')
-
-        if count != 0:
-            users = users[:count]
-
-        for u in users:
-            u.badges = Award.objects.filter(user=u, awardcourse__course__coursecohort__cohort=self).count()
-            if u.total is None:
-                u.total = 0
-        return users
-
-
-class CourseCohort(models.Model):
-    course = models.ForeignKey(Course, on_delete=models.CASCADE)
-    cohort = models.ForeignKey(Cohort, on_delete=models.CASCADE)
-
-    class Meta:
-        unique_together = ("course", "cohort")
-
-    
-class Participant(models.Model):
-    TEACHER = 'teacher'
-    STUDENT = 'student'
-    ROLE_TYPES = (
-        (TEACHER, 'Teacher'),
-        (STUDENT, 'Student'),
-    )
-    cohort = models.ForeignKey(Cohort, on_delete=models.CASCADE)
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    role = models.CharField(max_length=20, choices=ROLE_TYPES)
-
-    class Meta:
-        verbose_name = _('Participant')
-        verbose_name_plural = _('Participants')
-
-
-class Badge(models.Model):
-    ref = models.CharField(max_length=20)
-    name = models.TextField(blank=False)
-    description = models.TextField(blank=True)
-    default_icon = models.FileField(upload_to="badges")
-    points = models.IntegerField(default=100)
-    allow_multiple_awards = models.BooleanField(default=False)
-
-    class Meta:
-        verbose_name = _('Badge')
-        verbose_name_plural = _('Badges')
-
-    def __unicode__(self):
-        return self.description
-
-
-class Award(models.Model):
-    badge = models.ForeignKey(Badge, on_delete=models.CASCADE)
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    description = models.TextField(blank=False)
-    award_date = models.DateTimeField('date awarded', default=timezone.now)
-
-    class Meta:
-        verbose_name = _('Award')
-        verbose_name_plural = _('Awards')
-
-    def __unicode__(self):
-        return self.description
-
-    @staticmethod
-    def get_userawards(user, course=None):
-        awards = Award.objects.filter(user=user)
-        if course is not None:
-            awards = awards.filter(awardcourse__course=course)
-        return awards.count()
-
-    def _get_badge(self):
-        badge_icon = self.badge.default_icon
-        try:
-            icon = AwardCourse.objects.get(award=self)
-            if icon.course.badge_icon:
-                return icon.course.badge_icon
-        except AwardCourse.DoesNotExist:
-            pass
-        return badge_icon
-
-    badge_icon = property(_get_badge)
-
-
-class AwardCourse(models.Model):
-    award = models.ForeignKey(Award, on_delete=models.CASCADE)
-    course = models.ForeignKey(Course, on_delete=models.CASCADE)
-    course_version = models.BigIntegerField(default=0)
-
-
-class Points(models.Model):
-    POINT_TYPES = (
-        ('signup', 'Sign up'),
-        ('userquizattempt', 'Quiz attempt by user'),
-        ('firstattempt', 'First quiz attempt'),
-        ('firstattemptscore', 'First attempt score'),
-        ('firstattemptbonus', 'Bonus for first attempt score'),
-        ('quizattempt', 'Quiz attempt'),
-        ('quizcreated', 'Created quiz'),
-        ('activitycompleted', 'Activity completed'),
-        ('mediaplayed', 'Media played'),
-        ('badgeawarded', 'Badge awarded'),
-        ('coursedownloaded', 'Course downloaded'),
-    )
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    course = models.ForeignKey(Course, null=True, default=None, on_delete=models.SET_NULL)
-    points = models.IntegerField()
-    date = models.DateTimeField('date created', default=timezone.now)
-    description = models.TextField(blank=False)
-    data = models.TextField(blank=True)
-    type = models.CharField(max_length=20, choices=POINT_TYPES)
-
-    class Meta:
-        verbose_name = _('Points')
-        verbose_name_plural = _('Points')
-
-    def __unicode__(self):
-        return self.description
-
-    @staticmethod
-    def get_leaderboard(count=0, course=None):
-
-        from summary.models import UserCourseSummary, UserPointsSummary
-
-        if course is not None:
-            users = UserCourseSummary.objects.filter(course=course)
-            users_points = users.values('user').annotate(points=Sum('points'), badges=Sum('badges_achieved')).order_by('-points')
-        else:
-            users_points = UserPointsSummary.objects.all().values('user', 'points', 'badges').order_by('-points')
-
-        if count > 0:
-            users_points = users_points[:count]
-
-        leaderboard = []
-        for u in users_points:
-            user = User.objects.get(pk=u['user'])
-            user.badges = 0 if u['badges'] is None else u['badges']
-            user.total = 0 if u['points'] is None else u['points']
-            leaderboard.append(user)
-
-        return leaderboard
-    
-    @staticmethod
-    def get_userscore(user):
-        score = Points.objects.filter(user=user).aggregate(total=Sum('points'))
-        if score['total'] is None:
-            return 0
-        return score['total']
-
-    @staticmethod
-    def media_points(user, start_date=None, end_date=None, course=None):
-        results = Points.objects.filter(user=user, type='mediaplayed')
-        if start_date:
-            results = results.filter(date__gte=start_date)
-        if end_date:
-            results = results.filter(date__lte=end_date)
-        if course:
-            results = results.filter(course=course)
-        score = results.aggregate(total=Sum('points'))
-        if score['total'] is None:
-            return 0
-        return score['total']
-
-    @staticmethod
-    def page_points(user, start_date=None, end_date=None, course=None):
-        results = Points.objects.filter(user=user, type='activitycompleted')
-        if start_date:
-            results = results.filter(date__gte=start_date)
-        if end_date:
-            results = results.filter(date__lte=end_date)
-        if course:
-            results = results.filter(course=course)
-        score = results.aggregate(total=Sum('points'))
-        if score['total'] is None:
-            return 0
-        return score['total']
-
-    @staticmethod
-    def quiz_points(user, start_date=None, end_date=None, course=None):
-        results = Points.objects.filter(user=user).filter(Q(type='firstattempt') | Q(type='firstattemptscore') | Q(type='firstattemptbonus') | Q(type='quizattempt'))
-        if start_date:
-            results = results.filter(date__gte=start_date)
-        if end_date:
-            results = results.filter(date__lte=end_date)
-        if course:
-            results = results.filter(course=course)
-        score = results.aggregate(total=Sum('points'))
-        if score['total'] is None:
-            return 0
-        return score['total']
