+++ oppiamobile/settings.py
@@ -183,7 +183,7 @@ BADGE_AWARD_METHOD_ALL_QUIZZES = 'all quizzes'
 
 BADGE_AWARDING_METHOD = BADGE_AWARD_METHOD_ALL_ACTIVITIES
 
-OPPIA_GOOGLE_ANALYTICS_ENABLED = True
+OPPIA_GOOGLE_ANALYTICS_ENABLED = False
 OPPIA_GOOGLE_ANALYTICS_CODE = 'YOUR_GOOGLE_ANALYTICS_CODE'
 OPPIA_GOOGLE_ANALYTICS_DOMAIN = 'YOUR_DOMAIN'
