+++ oppiamobile/settings_secret.py.template
@@ -52,3 +52,9 @@ DEVICE_ADMIN_ENABLED = False
 # If the Device Admin is enabled, must provide a GCM ApiKey
 # https://console.firebase.google.com/project/_/settings/cloudmessaging/
 GCM_APIKEY = 'OPPIA_GOOGLEAPIKEY'
+
+
+OPPIA_ALLOW_SELF_REGISTRATION = True    # turns on/off ability for users to self register
+OPPIA_SHOW_GRAVATARS = True
+OPPIA_STAFF_ONLY_UPLOAD = True          # prevents anyone without is_staff status being able to upload courses,
+# setting to False allows any registered user to upload a course
